requirements.txt
flask==1.1.2
pillow==8.1.1
tensorflow==1.15.4
jupyter==1.0.0
torchvision==0.9.1
pillow==8.1.1

